requirements.txt
asgiref==3.2.3
boto3==1.12.0
Django==3.0.3
djangorestframework==3.11.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0
